services/user-service/pyproject.toml
[project]
name = "finans-tracker-user-service"
version = "0.1.0"
description = "User microservice for Finans Tracker"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.13.0",
    "bcrypt>=4.1.0",
    "python-jose[cryptography]>=3.3.0",
    "aio-pika>=9.4.0",
    "pydantic[email]>=2.0",
    "pydantic-settings>=2.0",
    "finans-tracker-contracts",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "httpx",
    "aiosqlite",
]

[tool.uv.sources]
finans-tracker-contracts = { path = "../shared/contracts" }

[tool.pytest.ini_options]
asyncio_mode = "auto"
